--- babel-hoist.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/babel-hoist.js')
