--- jsxstyle.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/jsxstyle.js')
