+++ lerna.json
@@ -1,0 +1,7 @@
+{
+  "lerna": "2.0.0-beta.32",
+  "packages": [
+    "packages/*"
+  ],
+  "version": "0.0.0"
+}
