+++ packages/babel-plugin-glamor/babel-hoist.js
