+++ packages/glamor-aphrodite/tests/index.js
@@ -1,0 +1,3 @@
+describe('aphrodite', () => {
+  it('takes an aphrodite stylesheet and applies it onto a dom')
+})
