+++ packages/glamor-raw-css/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "glamor-raw-css",
+  "version": "2.20.22",
+  "description": "inline css for component systems",
+  "main": "lib/index.js",
+  "typings": "index.d.ts",
+  "scripts": {
+    "parser": "pegjs src/spec.pegjs"
+  },
+  "author": "Sunil Pai <threepointone@gmail.com>",
+  "homepage": "https://github.com/threepointone/glamor",
+  "license": "MIT",
+  "repository": "https://github.com/threepointone/glamor",
+  "keywords": [
+    "css",
+    "inline"
+  ],
+  "dependencies": {
+    "object-assign": "^4.1.0",
+    "glamor": "*"
+  },
+  "devDependencies": {
+    "react": "^15.4.1",
+    "pegjs": "^0.10.0"
+  }
+}
