+++ packages/glamor-react/react.d.ts
