+++ packages/glamor/src/clean.js
