+++ packages/glamor/src/hash.js
