--- utils.js
@@ -1,2 +1,0 @@
-module.exports = require('./lib/utils.js'
-  )
